moved the dataset class to a new python file
